功能模块逻辑删除-改为物理删除

+++ vmes-server/src/main/java/com/xy/vmes/deecoop/system/controller/DictionaryController.java
@@ -387,26 +387,31 @@ public class DictionaryController {
      */
     @PostMapping("/dictionary/deleteDictionarys")
     public ResultModel deleteDictionarys()  throws Exception {
-
         logger.info("################dictionary/deleteDictionarys 执行开始 ################# ");
         Long startTime = System.currentTimeMillis();
-        HttpServletResponse response  = HttpUtils.currentResponse();
         ResultModel model = new ResultModel();
+
         PageData pd = HttpUtils.parsePageData();
-//        String dictionaryIds = pd.getString("ids");
-//        String[] ids = dictionaryIds.split(",");
-//        dictionaryService.updateToDisableByIds(ids);
         String ids = pd.getString("ids");
         if(StringUtils.isEmpty(ids)){
             model.putCode("1");
             model.putMsg("未勾选删除记录，请重新选择！");
             return model;
         }
+
         String id_str = StringUtil.stringTrimSpace(ids);
+        String msgStr = dictionaryService.checkDeleteDictionaryByIds(id_str);
+        if (msgStr != null && msgStr.trim().length() > 0) {
+            model.putCode(Integer.valueOf(1));
+            model.putMsg(msgStr);
+            return model;
+        }
+
         String[] id_arry = id_str.split(",");
         if(id_arry.length>0){
             dictionaryService.deleteByIds(id_arry);
         }
+
         Long endTime = System.currentTimeMillis();
         logger.info("################dictionary/deleteDictionarys 执行结束 总耗时"+(endTime-startTime)+"ms ################# ");
         return model;
